requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.29.0
gast==0.3.3
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.10.1
imutils==0.5.4
joblib==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==2.0.0
oauthlib==3.1.1
opencv-python==4.3.0.36
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pipreqs==0.4.11
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PyPrind==2.11.3
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0
